pyproject.toml
[project]
name = "classifier-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["anthropic>=0.76.0", "dspy>=3.1.0", "python-dotenv>=1.2.1"]

[dependency-groups]
dev = [
    "langchain-anthropic>=1.3.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ragas>=0.4.3",
    "ruff>=0.14.11",
    "ty>=0.0.12",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/classifier_demo"]
